maintenance_standards/requirements.txt
# 基础依赖
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# 文档处理
python-docx>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# 知识图谱和NLP
langchain>=0.1.0
neo4j>=5.10.0
spacy>=3.7.0

# Web界面
gradio>=4.0.0
pyvis>=0.3.0
networkx>=3.1

# 开发工具
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0